split off TestTimeGrouper class
